+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 All notable changes to the project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [gtk-fortran-extra development]
+## [gtk-fortran-extra 1.2.0] 2023-04-22
 
 ### Added
+- `saville_code/`: encodes a text using Peter Saville's color code.
 - `unknown_pleasures/` can now also be built with the Fortran Package Manager fpm.
 - `unknown_pleasures/`: Cairo text was added on the figure. It can be deactivated with a boolean flag in the code.
 - `parallel_app/` demonstrates how you can use parallel Fortran 2008 and 2018 features (coarrays, collective routines...) in a gtk-fortran application.